+++ sam/models/lasso_model.py
@@ -1,5 +1,5 @@
-import os
-from typing import Callable, Sequence, Tuple, Union
+from pathlib import Path
+from typing import Callable, Sequence, Tuple, Union, Any
 
 import numpy as np
 import pandas as pd
@@ -191,43 +191,16 @@ class LassoTimeseriesRegressor(BaseTimeseriesRegressor):
         else:
             return prediction
 
-    def dump(self, foldername: str, prefix: str = "model") -> None:
-        """Save a model to disk
+    def dump_parameters(self, foldername: str, prefix: str = "model") -> None:
+        import cloudpickle
 
-        This abstract method needs to be implemented by any class inheriting from
-        SamQuantileRegressor. This function dumps the SAM model to disk.
+        with open(Path(foldername) / f"{prefix}_params.pkl", "wb") as f:
+            cloudpickle.dump(self.model_, f)
 
-        Parameters
-        ----------
-        foldername : str
-            The folder location where to save the model
-        prefix : str, optional
-           The prefix used in the filename, by default "model"
-        """
-        import joblib
+    @staticmethod
+    def load_parameters(obj, foldername: str, prefix: str = "model") -> Any:
+        import cloudpickle
 
-        if not os.path.exists(foldername):
-            os.makedirs(foldername)
-        joblib.dump(self, os.path.join(foldername, f"{prefix}.pkl"))
-
-    @classmethod
-    def load(cls, foldername, prefix="model") -> Callable:
-        """Load a model from disk
-
-        This abstract method needs to be implemented by any class inheriting from
-        SamQuantileRegressor. This function loads a SAM model from disk.
-
-        Parameters
-        ----------
-        foldername : str
-            The folder location where the model is stored
-        prefix : str, optional
-           The prefix used in the filename, by default "model"
-
-        Returns
-        -------
-        The SAM model that has been loaded from disk
-        """
-        import joblib
-
-        return joblib.load(os.path.join(foldername, f"{prefix}.pkl"))
+        with open(Path(foldername) / f"{prefix}_params.pkl", "rb") as f:
+            model = cloudpickle.load(f)
+        return model